HomeActivity -> HomeFragment. HomeFragment and Home merged to form HomeFragment. Extra layout files removed. Layout file renamed to follow convention.
