+++ app/build.gradle
@@ -21,8 +21,8 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     //noinspection GradleCompatible
-    implementation 'com.android.support:appcompat-v7:28.0.0-rc01'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'com.android.support:appcompat-v7:28.0.0-alpha1'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     implementation 'com.google.android.gms:play-services-auth:16.0.0'
     implementation 'com.android.volley:volley:1.1.0'
     implementation 'com.google.code.gson:gson:2.8.2'
@@ -38,11 +38,8 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.8.0'
     implementation 'de.hdodenhof:circleimageview:2.2.0'
 
-    //Recycler View
-    implementation 'com.android.support:recyclerview-v7:28.0.+'
-
     //CardView
-    implementation 'com.android.support:cardview-v7:28.0.0-alpha1'
+    implementation 'com.android.support:cardview-v7:28.0.0-rc02'
 
 }
 allprojects {
