--- app/src/main/java/com/example/lenovo/engineer/Home.java
@@ -1,73 +1,0 @@
-package com.example.lenovo.engineer;
-
-import android.os.Bundle;
-import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-
-import java.util.ArrayList;
-import java.util.Timer;
-import java.util.TimerTask;
-
-import me.relex.circleindicator.CircleIndicator;
-
-public class Home extends Fragment {
-    private static ViewPager mPager;
-    private static int currentPage = 0;
-    private static final Integer[] Images = {R.drawable.pic1, R.drawable.pic2, R.drawable.pic3};
-    private ArrayList<Integer> ImageArray = new ArrayList<Integer>();
-
-    @Nullable
-    @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        //returning our layout file
-        //change R.layout.yourlayoutfilename for each of your fragments
-        return inflater.inflate(R.layout.home_resource_1, container, false);
-
-    }
-
-    @Override
-    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        //you can set the title for your toolbar here for different fragments different titles
-        getActivity().setTitle("Engineer");
-
-        initSlider();
-    }
-
-    private void initSlider() {
-        for (int i = 0; i < Images.length; i++)
-            ImageArray.add(Images[i]);
-
-        mPager = (ViewPager) getView().findViewById(R.id.pager);
-
-        mPager.setAdapter(new ImageAdapter(getContext(), ImageArray));
-        CircleIndicator indicator = getView().findViewById(R.id.indicator);
-        indicator.setViewPager(mPager);
-
-        // Auto start of viewpager
-        final Handler handler = new Handler();
-        final Runnable Update = new Runnable() {
-            public void run() {
-                if (currentPage == Images.length) {
-                    currentPage = 0;
-                }
-                mPager.setCurrentItem(currentPage++, true);
-            }
-        };
-
-        Timer swipeTimer = new Timer();
-        swipeTimer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                handler.post(Update);
-            }
-        }, 2500, 2500);
-    }
-}
