--- app/src/main/java/com/example/lenovo/engineer/HomeActivity.java
@@ -1,306 +1,0 @@
-package com.example.lenovo.engineer;
-
-import android.content.Intent;
-import android.content.res.Resources;
-import android.graphics.Rect;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.AppBarLayout;
-import android.support.design.widget.CollapsingToolbarLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
-import android.util.TypedValue;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.request.RequestOptions;
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class HomeActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
-
-    private RecyclerView recyclerView;
-    private CommAdapter adapter;
-    private List<CommitteeDet> committeeDetList;
-    private static final String TAG = "HomeActivity";
-    public HomeActivity()
-    {
-
-    }
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.home_activity_main);
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.home_toolbar);
-        setSupportActionBar(toolbar);
-        toolbar.showOverflowMenu();
-
-        //Adding Drawer to layout
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.home_drawer_layout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.addDrawerListener(toggle);
-        toggle.syncState();
-
-        //Adding Fab button
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(HomeActivity.this, FavSchedule.class);
-                startActivity(intent);
-            }
-        });
-
-        //
-
-        NavigationView navigationView = (NavigationView) findViewById(R.id.home_nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
-        // Removing this will cause the navigation view to not respond to click events
-        navigationView.bringToFront();
-
-        View header = navigationView.getHeaderView(0);
-        GoogleSignInAccount account = GoogleSignInHelper.getInstance().getAccount();
-
-        View navView;
-        navView = header.findViewById(R.id.name);
-        ((TextView) navView).setText(account.getDisplayName());
-        navView = header.findViewById(R.id.email);
-        ((TextView) navView).setText(account.getEmail());
-        navView = header.findViewById(R.id.profilePic);
-        Uri imguri = account.getPhotoUrl();
-
-        RequestOptions options = new RequestOptions()
-                .centerCrop()
-                .placeholder(R.mipmap.ic_launcher_round)
-                .error(R.mipmap.ic_launcher_round);
-
-        Glide.with(this).load(imguri).apply(options).into(((ImageView) navView));
-        //End of drawer
-
-
-        initCollapsingToolbar();
-        recyclerView=(RecyclerView) findViewById(R.id.home_recycler_view);
-        committeeDetList=new ArrayList<>();
-        adapter=new CommAdapter(this,committeeDetList);
-
-        RecyclerView.LayoutManager mLayoutManager= new GridLayoutManager(this,2);
-        recyclerView.setLayoutManager(mLayoutManager);
-        recyclerView.addItemDecoration(new GridSpacingItemDecoration(2, dpToPx(10), true));
-        recyclerView.setItemAnimator(new DefaultItemAnimator());
-        prepCommitte();
-        recyclerView.setAdapter(adapter);
-        try {
-            Glide.with(this).load(R.drawable.logo).into((ImageView) findViewById(R.id.backdrop));
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-
-
-    }
-    private void initCollapsingToolbar() {
-        final CollapsingToolbarLayout collapsingToolbar =
-                (CollapsingToolbarLayout) findViewById(R.id.collapsing_toolbar);
-        collapsingToolbar.setTitle(" ");
-        AppBarLayout appBarLayout = (AppBarLayout) findViewById(R.id.appbar);
-        appBarLayout.setExpanded(true);
-
-        // hiding & showing the title when toolbar expanded & collapsed
-        appBarLayout.addOnOffsetChangedListener(new AppBarLayout.OnOffsetChangedListener() {
-            boolean isShow = false;
-            int scrollRange = -1;
-
-            @Override
-            public void onOffsetChanged(AppBarLayout appBarLayout, int verticalOffset) {
-                if (scrollRange == -1) {
-                    scrollRange = appBarLayout.getTotalScrollRange();
-                }
-                if (scrollRange + verticalOffset == 0) {
-                    collapsingToolbar.setTitle(getString(R.string.app_name));
-                    isShow = true;
-                } else if (isShow) {
-                    collapsingToolbar.setTitle(" ");
-                    isShow = false;
-                }
-            }
-        });
-    }
-
-
-
-    private void prepCommitte() {
-        int[] covers = new int[]
-                {
-                        R.drawable.tronix,
-                        R.drawable.mechanical,
-                        R.drawable.meta,
-                        R.drawable.comps,
-                        R.drawable.chemical,
-                        R.drawable.mining,
-                        R.drawable.civil,
-                        R.drawable.technites,
-                };
-        CommitteeDet a= new CommitteeDet(getString(R.string.Tronix),covers[0]);
-        committeeDetList.add(a);
-        CommitteeDet c= new CommitteeDet(getString(R.string.Mech),covers[1]);
-        CommitteeDet d= new CommitteeDet(getString(R.string.Meta),covers[2]);
-        CommitteeDet e= new CommitteeDet(getString(R.string.Comp),covers[3]);
-        CommitteeDet f= new CommitteeDet(getString(R.string.Chem),covers[4]);
-        CommitteeDet g= new CommitteeDet(getString(R.string.Mining),covers[5]);
-        CommitteeDet h= new CommitteeDet(getString(R.string.Civil),covers[6]);
-        a= new CommitteeDet(getString(R.string.Technites),covers[7]);
-        committeeDetList.add(a);
-
-
-
-        committeeDetList.add(c);
-        committeeDetList.add(d);
-        committeeDetList.add(e);
-        committeeDetList.add(f);
-        committeeDetList.add(g);
-        committeeDetList.add(h);
-
-        adapter.notifyDataSetChanged();
-    }
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.main_menu, menu);
-        return true;
-    }
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        Log.d(TAG, "Menu item clicked.");
-        // Handle navigation view item clicks here.
-        int id = item.getItemId();
-        Fragment fragment = null;
-
-        if (id == R.id.Home) {
-
-        } else if (id == R.id.nav_schedule) {
-           // fragment = new ScheduleFragment();
-            Intent intent= new Intent(HomeActivity.this,main_menu.class);
-            startActivity(intent);
-            finish();
-
-        } else if (id == R.id.nav_maps) {
-            fragment = new MapsFragment();
-
-        } else if (id == R.id.nav_slideshow) {
-
-        } else if (id == R.id.nav_manage) {
-
-        } else if (id == R.id.nav_share) {
-
-        } else if (id == R.id.nav_send) {
-
-        } else if (id == R.id.logout) {
-            GoogleSignInHelper.getInstance().getClient().signOut();
-            Log.d(TAG, "Logout Successful");
-            finish();
-            startActivity(new Intent(this, MainActivity.class));
-        }
-
-        if (fragment != null) {
-            FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
-            ft.replace(R.id.home_relative_layout, fragment);
-            ft.commit();
-
-        }
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.home_drawer_layout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
-
-    }
-
-
-    /**
-     * RecyclerView item decoration - give equal margin around grid item
-     */
-
-    public class GridSpacingItemDecoration extends RecyclerView.ItemDecoration {
-
-        private int spanCount;
-        private int spacing;
-        private boolean includeEdge;
-
-        public GridSpacingItemDecoration(int spanCount, int spacing, boolean includeEdge) {
-            this.spanCount = spanCount;
-            this.spacing = spacing;
-            this.includeEdge = includeEdge;
-        }
-
-        @Override
-        public void getItemOffsets(Rect outRect, View view, RecyclerView parent, RecyclerView.State state) {
-            int position = parent.getChildAdapterPosition(view); // item position
-            int column = position % spanCount; // item column
-
-            if (includeEdge) {
-                outRect.left = spacing - column * spacing / spanCount; // spacing - column * ((1f / spanCount) * spacing)
-                outRect.right = (column + 1) * spacing / spanCount; // (column + 1) * ((1f / spanCount) * spacing)
-
-                if (position < spanCount) { // top edge
-                    outRect.top = spacing;
-                }
-                outRect.bottom = spacing; // item bottom
-            } else {
-                outRect.left = column * spacing / spanCount; // column * ((1f / spanCount) * spacing)
-                outRect.right = spacing - (column + 1) * spacing / spanCount; // spacing - (column + 1) * ((1f /    spanCount) * spacing)
-                if (position >= spanCount) {
-                    outRect.top = spacing; // item top
-                }
-            }
-        }
-    }
-    @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.home_drawer_layout);
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    private int dpToPx(int dp) {
-        Resources r = getResources();
-        return Math.round(TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, r.getDisplayMetrics()));
-    }
-}
