--- app/src/main/res/layout/home_content_main.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/home_relative_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/viewBg"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.lenovo.engineer.HomeActivity"
-    tools:showIn="@layout/home_activity_main">
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/home_recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:clipToPadding="false"
-        android:scrollbars="vertical" />
-
-</RelativeLayout>
